requirements.txt
numpy==1.26.3
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.2.0
pygrabber==0.2
PyMySQL==1.1.0
PyQt6==6.6.1
PyQt6_sip==13.6.0
pyserial==3.5
python_escpos==3.0
pyzbar==0.1.9
Requests==2.31.0
